refactor(config,launcher,ui): drop client-config/Network.kt; move NetworkState into network/
Conduit Phase 3 left Network.kt with only two timeout constants and
NetworkState was sitting at the top of `hivens.launcher` package while
ChannelRouter (its principal caller) lives one level deeper in
`hivens.launcher.network` -- the child package reaching up to grab a
sibling concept. Finish the cleanup:

- Add `connectTimeoutMs` and `readTimeoutMs` fields to
  ServerProtocolConfig with the same defaults the old constants held.
  Mirror operators with slower upstreams get a knob; the config file
  override path already exists.
- Migrate the eight OkHttpClient builder sites in Modules.kt to read
  from cfg.connectTimeoutMs / cfg.readTimeoutMs.
- Delete `client-config/Network.kt` -- it had no other surface left.
- Move `NetworkState` from `hivens.launcher` to `hivens.launcher.network`
  (sibling of ChannelRouter, ServerProtocolConfig). Renames the package
  declaration on NetworkState.kt + NetworkStateTest.kt; the seven import
  sites across ui/RightPanel, ui/Main, ui/AppLayout, ui/screens/Settings
  and Dashboard, launcher/di/Modules, ChannelRouter (and the test) all
  point at the new location.
- Update KDoc cross-references in HttpClientProvider.kt and Modules.kt
  that pointed at the now-deleted `hivens.config.Network`.

`Network` constant taxonomy is fully retired. Anything network-related
that needs configuration now lives on ServerProtocolConfig; cross-
cutting infra constants live nowhere because there aren't any to be.
