+++ client-launcher/src/main/kotlin/hivens/launcher/network/ChannelRouter.kt
@@ -1,7 +1,7 @@
 package hivens.launcher.network
 
 import hivens.core.diag.ActionRing
-import hivens.launcher.NetworkState
+import hivens.launcher.network.NetworkState
 import io.ktor.client.HttpClient
 import org.slf4j.LoggerFactory
 import java.io.IOException
