+++ client-launcher/src/main/kotlin/hivens/launcher/network/NetworkState.kt
@@ -1,4 +1,4 @@
-package hivens.launcher
+package hivens.launcher.network
 
 import hivens.core.security.SslBypassEntry
 import kotlinx.serialization.decodeFromString
